PyTorch distributed num-workers>0 fix (#252)
* seperating dataset code in dedicated script

* remove unpickable dataset_info

* set default prefetch factor for num-workers>0

* better doc

* add name to Namespace

* tensorboard
